+++ cli/bag/bag_commands.h
@@ -31,10 +31,10 @@ int bag_info(const std::string& path);
 
 int bag_record(const std::string& path, const std::vector<std::string>& urls, const std::string& tag_name,
                const std::string& filter, bool black_mode, bool native_mode, double duration, double wait_time,
-               bool compress, bool force, int64_t max_row_count, double max_bytes_size, bool enable_limit,
-               bool split_name_by_time, double split_by_size, int64_t split_by_time, int64_t max_split_count, bool deft,
-               double max_packet_size, bool wal_mode, double cache_size, bool sync_mode,
-               const std::vector<std::string>& ignore_compress, const std::string& plugin_name);
+               bool compress, bool enable_chunk_crc, bool force, int64_t max_row_count, double max_bytes_size,
+               bool enable_limit, bool split_name_by_time, double split_by_size, int64_t split_by_time,
+               int64_t max_split_count, bool deft, double max_packet_size, bool wal_mode, double cache_size,
+               bool sync_mode, const std::vector<std::string>& ignore_compress, const std::string& plugin_name);
 
 int bag_play(const std::string& path, const std::vector<std::string>& urls, const std::string& filter, bool black_mode,
              bool native_mode, bool auto_pause, const std::vector<int>& actions, int64_t begin_time, int64_t end_time,
@@ -44,15 +44,17 @@ int bag_play(const std::string& path, const std::vector<std::string>& urls, cons
 int bag_clone(const std::string& source_path, const std::string& target_path, const std::vector<std::string>& urls,
               const std::string& tag_name, const std::string& filter, bool black_mode, const std::vector<int>& actions,
               int64_t begin_time, int64_t end_time, bool has_clock_begin_time, bool has_clock_end_time, bool compress,
-              bool split_name_by_time, double split_by_size, int64_t split_by_time, bool force, bool wal_mode,
-              double cache_size, const std::vector<std::string>& ignore_compress, const std::string& plugin_name);
+              bool enable_chunk_crc, bool split_name_by_time, double split_by_size, int64_t split_by_time, bool force,
+              bool wal_mode, double cache_size, const std::vector<std::string>& ignore_compress,
+              const std::string& plugin_name);
 
 int bag_merge(const std::vector<std::string>& source_paths, const std::string& target_path,
               const std::vector<std::string>& urls, const std::string& tag_name, const std::string& filter,
               bool black_mode, const std::vector<int>& actions, int64_t begin_time, int64_t end_time,
-              bool has_clock_begin_time, bool has_clock_end_time, bool compress, bool split_name_by_time,
-              double split_by_size, int64_t split_by_time, bool force, bool wal_mode, double cache_size,
-              const std::vector<std::string>& ignore_compress, const std::string& plugin_name, double check_gap);
+              bool has_clock_begin_time, bool has_clock_end_time, bool compress, bool enable_chunk_crc,
+              bool split_name_by_time, double split_by_size, int64_t split_by_time, bool force, bool wal_mode,
+              double cache_size, const std::vector<std::string>& ignore_compress, const std::string& plugin_name,
+              double check_gap);
 
 int bag_check(const std::string& path);